agent/mcp-agent-auth/mcp_snowflake/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp_snowflake"
version = "0.1.0"
description = "MCP Tool for Websearch"
requires-python = "==3.13.3"
dependencies = [
    "langchain-community ~=0.3.16",
    "langchain-core ~=0.3.33",
    "langchain-groq ~=0.2.1",
    "langchain-google-genai ~=2.0.11",
    "langgraph ~=0.4.5",
    "langgraph-checkpoint-mongodb ~=0.1.3",
    "langgraph-checkpoint-postgres ~=2.0.13",
    "langgraph-checkpoint-sqlite ~=2.0.1",
    "pydantic ~=2.11.4",
    "uvicorn ~=0.32.1",
    "fastapi ~=0.115.5",
    "resend==2.8.0",
    "langchain-mcp-adapters==0.1.0",
    "streamlit ~=1.45.1",
    "scikit-learn ~=1.6.1",
    "snowflake-connector-python==3.15.0",
    "fastmcp==2.8.1"
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "mypy",
    "pytest",
    "pytest-cov",
]

[tool.hatch.build.targets.wheel]
packages = ["mcp_snowflake"]
